moving python code base into module and tests folder
