+++ pd2ppt/__init__.py
@@ -1,0 +1,2 @@
+from .pd2ppt import df_to_powerpoint
+from .pd2ppt import df_to_table
